backend/pyproject.toml
[project]
name = "infrajet-backend"
version = "0.1.4"
description = "Infrajet Backend API with FastAPI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Core
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.4.2",
    "pydantic-settings>=2.0.3",
    # Database
    "sqlalchemy[asyncio]>=2.0.23",
    "alembic>=1.12.1",
    "asyncpg>=0.28.0",
    # Email
    "aiosmtplib>=2.0.1",
    "email-validator>=2.1.0",
    # Tree-sitter and HCL parsing
    "tree-sitter>=0.20.0",
    "python-hcl2>=4.0.0",
    "pyyaml>=6.0",
    "loguru>=0.7.3",
    "anthropic>=0.64.0",
    "redis>=4.0.0",
    "faiss-cpu>=1.12.0",
    "psycopg2-binary>=2.9.10",
    "pgvector>=0.4.1",
    "psutil>=7.0.0",
    "prometheus-client>=0.19.0",
    "azure-storage-file-share>=12.22.0",
    "aiohttp>=3.12.15",
    # Azure Entra ID integration
    "msal>=1.24.0",
    "azure-identity>=1.15.0",
    # GitHub integration and token encryption
    "cryptography>=41.0.0",
    "itsdangerous>=2.2.0",
    # Socket.IO for real-time communication
    "python-socketio>=5.8.0",
    "python-socketio[asyncio-client]>=5.8.0",
]

[project.optional-dependencies]
dev = [
    # Development
    "debugpy>=1.8.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.1",
    
    # Code Quality
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "pylint>=3.0.2",
    
    # Documentation
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.1",
]

# Tree-sitter language bindings (optional for enhanced parsing)
tree-sitter = [
    # Alternative HCL parsers that are available on PyPI
    "lark>=1.1.0",  # Alternative parser for complex syntax
    "tree-sitter-languages>=0.1.0",  # Pre-built language bindings including HCL and Terraform
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
